requirements.txt
python-decouple
aiogram
redis
apscheduler
